refactor: ai-dto 파일구조 분리 (typecast & gpt)
