+++ src/main/java/com/kuit/agarang/domain/ai/model/dto/typecast/TypecastMessageRequest.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.ai.model.dto;
+package com.kuit.agarang.domain.ai.model.dto.typecast;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class MessageRequest {
+public class TypecastMessageRequest {
   // TODO : actor 정보 추가 (남아 or 여아)
   private String text;
 }
