+++ src/main/java/com/kuit/agarang/domain/ai/model/dto/typecast/TypecastRequest.java
@@ -1,4 +1,4 @@
-package com.kuit.agarang.domain.ai.model.dto;
+package com.kuit.agarang.domain.ai.model.dto.typecast;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
@@ -39,7 +39,7 @@ public class TypecastRequest {
     this.xapiAudioFormat = xapiAudioFormat;
   }
 
-  public static TypecastRequest create(MessageRequest request, String actorId) {
+  public static TypecastRequest create(TypecastMessageRequest request, String actorId) {
     return TypecastRequest.builder()
       .actorId(actorId)
       .text(request.getText())
